chore(sdk): Split typedefs into separate modules
