+++ packages/grafbase-sdk/src/auth/jwks.ts
@@ -1,8 +1,8 @@
-import { RequireExactlyOne } from "type-fest"
+import { RequireExactlyOne } from 'type-fest'
 
 export type JWKSParams = {
-  issuer?: string,
-  jwksEndpoint?: string,
+  issuer?: string
+  jwksEndpoint?: string
   clientId?: string
   groupsClaim?: string
 }
@@ -13,7 +13,9 @@ export class JWKSAuth {
   clientId?: string
   groupsClaim?: string
 
-  constructor(params: RequireExactlyOne<JWKSParams, 'issuer' | 'jwksEndpoint'>) {
+  constructor(
+    params: RequireExactlyOne<JWKSParams, 'issuer' | 'jwksEndpoint'>
+  ) {
     this.issuer = params.issuer
     this.jwksEndpoint = params.jwksEndpoint
     this.clientId = params.clientId
@@ -21,10 +23,14 @@ export class JWKSAuth {
   }
 
   public toString(): string {
-    const issuer = this.issuer ? `issuer: "${this.issuer}"` : ""
-    const jwksEndpoint = this.jwksEndpoint ? `jwksEndpoint: "${this.jwksEndpoint}"` : ""
-    const clientId = this.clientId ? `, clientId: "${this.clientId}"` : ""
-    const groupsClaim = this.groupsClaim ? `, groupsClaim: "${this.groupsClaim}"` : ""
+    const issuer = this.issuer ? `issuer: "${this.issuer}"` : ''
+    const jwksEndpoint = this.jwksEndpoint
+      ? `jwksEndpoint: "${this.jwksEndpoint}"`
+      : ''
+    const clientId = this.clientId ? `, clientId: "${this.clientId}"` : ''
+    const groupsClaim = this.groupsClaim
+      ? `, groupsClaim: "${this.groupsClaim}"`
+      : ''
 
     return `{ type: jwks, ${issuer}${jwksEndpoint}${clientId}${groupsClaim} }`
   }
