+++ packages/grafbase-sdk/src/interface.ts
@@ -1,6 +1,16 @@
 import { Field } from './field'
 import { ListDefinition } from './field/list'
-import { ScalarDefinition } from './field/typedefs'
+import { ScalarDefinition } from './typedefs/scalar'
+
+/**
+ * A collection of fields in an interface.
+ */
+export type InterfaceFields = Record<string, InterfaceFieldShape>
+
+/**
+ * A combination of classes a field in an interface can be.
+ */
+export type InterfaceFieldShape = ScalarDefinition | ListDefinition
 
 export class Interface {
   name: string
@@ -11,10 +21,7 @@ export class Interface {
     this.fields = []
   }
 
-  public field(
-    name: string,
-    definition: ScalarDefinition | ListDefinition
-  ): Interface {
+  public field(name: string, definition: InterfaceFieldShape): Interface {
     this.fields.push(new Field(name, definition))
 
     return this
@@ -22,9 +29,7 @@ export class Interface {
 
   public toString(): string {
     const header = `interface ${this.name} {`
-
     const fields = this.fields.map((field) => `  ${field}`).join('\n')
-
     const footer = '}'
 
     return `${header}\n${fields}\n${footer}`
