+++ packages/grafbase-sdk/src/typedefs/length-limited-string.ts
@@ -1,0 +1,59 @@
+import { RequireAtLeastOne } from 'type-fest'
+import { FieldType } from '../field/typedefs'
+import { UniqueDefinition } from './unique'
+import { DefaultDefinition } from './default'
+import { Enum } from '../enum'
+import { StringDefinition } from './scalar'
+import { SearchDefinition } from './search'
+
+export interface FieldLength {
+  min?: number
+  max?: number
+}
+
+export class LengthLimitedStringDefinition {
+  fieldLength: RequireAtLeastOne<FieldLength, 'min' | 'max'>
+  scalar: StringDefinition
+
+  constructor(
+    scalar: StringDefinition,
+    fieldLength: RequireAtLeastOne<FieldLength, 'min' | 'max'>
+  ) {
+    this.fieldLength = fieldLength
+    this.scalar = scalar
+  }
+
+  public unique(scope?: string[]): UniqueDefinition {
+    return new UniqueDefinition(this, scope)
+  }
+
+  public search(): SearchDefinition {
+    return new SearchDefinition(this)
+  }
+
+  public default(val: string): DefaultDefinition {
+    return new DefaultDefinition(this, val)
+  }
+
+  public optional(): LengthLimitedStringDefinition {
+    this.scalar.optional()
+
+    return this
+  }
+
+  fieldTypeVal(): FieldType | Enum {
+    return this.scalar.fieldType
+  }
+
+  public toString(): string {
+    const length = this.fieldLength
+
+    if (length.min != null && length.max != null) {
+      return `${this.scalar} @length(min: ${length.min}, max: ${length.max})`
+    } else if (length.min != null) {
+      return `${this.scalar} @length(min: ${length.min})`
+    } else {
+      return `${this.scalar} @length(max: ${length.max})`
+    }
+  }
+}
